--- packages/components/src/DashboardLayout/DashboardLayoutStoryTemplate.tsx
@@ -1,11 +1,0 @@
-import { type ComponentStory } from '@storybook/react';
-
-import { DashboardLayoutStory } from './DashboardLayoutStory';
-
-export const DashboardLayoutStoryTemplate: ComponentStory<
-  typeof DashboardLayoutStory
-> = (args) => {
-  return <DashboardLayoutStory {...args} />;
-};
-
-export default DashboardLayoutStoryTemplate;
