+++ packages/components/src/DashboardLayout/Main/Main.tsx
@@ -1,6 +1,6 @@
 import { type ReactNode, forwardRef } from 'react';
 
-import { MainRoot } from './styled';
+import { MainRoot } from './styles';
 
 export type MainProps = { children: ReactNode };
